chore(clean-types): starts moving some types into 'common'
