+++ runem/runem.py
@@ -45,14 +45,13 @@ from runem.job_execute import job_execute
 from runem.job_filter import filter_jobs
 from runem.log import error, log, warn
 from runem.report import report_on_run
+from runem.types_wip.common import OrderedPhases, PhaseName
 from runem.types_wip.types import (
     Config,
     FilePathListLookup,
     HookName,
     Jobs,
-    OrderedPhases,
     PhaseGroupedJobs,
-    PhaseName,
 )
 from runem.types_wip.types_jobs import (
     JobReturn,
