+++ runem/types_wip/common.py
@@ -1,0 +1,12 @@
+import typing
+
+# meta-data types
+JobName = str
+JobTag = str
+JobNames = typing.Set[JobName]
+JobPhases = typing.Set[str]
+JobTags = typing.Set[JobTag]
+PhaseName = str
+OrderedPhases = typing.Tuple[PhaseName, ...]
+FilePathSerialise = str
+FilePathList = typing.List[FilePathSerialise]
