+++ runem/types_wip/types_jobs.py
@@ -31,7 +31,8 @@ from datetime import timedelta
 
 from typing_extensions import Unpack
 
-from runem.types_wip.types import FilePathList, JobConfig, Options, PhaseName
+from runem.types_wip.common import FilePathList, PhaseName
+from runem.types_wip.types import JobConfig, Options
 
 ReportName = str
 ReportUrl = typing.Union[str, pathlib.Path]
